Add haml and slim to Ruby language files
I don't know if this is a good idea, since technically they aren't ruby files, but functions used in them can only be from ruby files, some of which may be helpers defined in the project. This works for me but I'm not sure if it's generally a good idea, but it seems like if js and jsx can share a language definition, perhaps so can ruby and its templating languages?