Add `dumb-jump-selector` for `helm`
This commit adds an additional option for `dumb-jumb-selector` for using
Helm to browse multi-choice jumps. A `if` clause has been changed to a
`cond` to accomodate the additional rule in a more elegant manner, but
the code is otherwise fairly similar to work done for `ivy` support.

Tests are forthcoming.