Project root detection: Ignore sub-folders using '.dumbjumpignore' (#36)
If a project has a ".dumbjumpignore" file it won't be detected as a project root even though it has
another project root denoter, like ".git" etc. This makes it possible to have a fallback top project
so searches can lookup things from other sub-projects.

Added a test case to verify it.