Sort non-current file matches higher to the top
Sort non-current files by path length so the nearest file is more likely to be sorted higher to the
top. Also sorts by line number for sanity.

When not preferring external, i.e. normal procedure, it sorts the non-current files so that the
order of the current file matches are kept intact. When preferring external matches it sorts those
external matches this way.