Sort matches better and correctly filter current file (#119)
* Match relative current file, too

In some cases the results will not be absolute but relative and the "current file" filters must
match in both cases.

* Test relative results are matched against absolute paths

* Sort non-current file matches higher to the top

Sort non-current files by path length so the nearest file is more likely to be sorted higher to the
top. Also sorts by line number for sanity.

When not preferring external, i.e. normal procedure, it sorts the non-current files so that the
order of the current file matches are kept intact. When preferring external matches it sorts those
external matches this way.

* Fix for relative current file detection when using testing paths

That is, if the paths doesn't exist they won't match against `default-directory'.