#150 enable lexical bindings (#152)
It's been observed that tests fail on GNU Emacs 24.3 if lexical-bindings are
enabled. There could be various reasons for this, but most obvious is
that Emacs 24.3 had `byte-compile-add-to-list` function, which was used
as byte-compiled version of `add-to-list`:
https://github.com/emacs-mirror/emacs/commit/0b31660d3c10a0f8e243dd67bd0ecaf2c847d1e6#diff-38337c5e732c14fa599af4a0f6e53f18L4272

ert-runner compiles files before the test run, so tests used the
compiled version.

This commit rewrites functions that depended on add-to-list. Closes #150