Add basic support for jumping to functions in SQL files
This is incredibly basic, and adds support for syntax generally defined by the SQL standard, with
provisions for syntax exposed by popular database engines, such as MySQL, PostgreSQL, MS-SQL etc.

Both upper-case and lower-case definitions will match. A better solution would'be been to allow for
case-insensitive matching, but mixed case within the same word should be rare enough to not be a
problem at this stage.