Simplify helm usage and fix persistent action
Before this commit, dumb-jump-prompt-user-for-choice would provide helm
only with the formatted choices.  The persistent action would need to parse
the selected choice, which did not always work properly because the path
may be truncated in the formatted choice.

* dumb-jump.el (dumb-jump-prompt-user-for-choice): Instead of invoking helm
and then invoking dumb-jump-to-selected on the result, invoke helm with a
default action that invokes dumb-jump-result-follow.
Pass in the actual results along with the formatted choices so that actions
have actual results to work with.

* dumb-jump.el (dumb-jump-helm-persist-action): Expect the actual result
rather than the formatted choice.
Fix the doc string.

* test/dumb-jump-test.el
(dumb-jump-prompt-user-for-choice-correct-helm-test,
dumb-jump-prompt-user-for-choice-correct-helm-persistent-action-test):
Update tests.