Jump to definition even though it took long to find it
The search process is not interrupted after `dumb-jump-max-find-time`
seconds but continues until all relevant files are searched for. Because of
that, it seams to be a waste to just print a message that the search took long
and do not jump to the found definition, if any.

This commit ensures that if any definition is found, we jump to it besides
printing the warning about the long search time.